chore(trie): replace ParallelStateRoot with AsyncStateRoot
