+++ crates/trie/parallel/Cargo.toml
@@ -32,12 +32,9 @@ tracing.workspace = true
 thiserror.workspace = true
 derive_more.workspace = true
 
-# `async` feature
-tokio = { workspace = true, optional = true, default-features = false }
-itertools = { workspace = true, optional = true }
-
 # `parallel` feature
 rayon = { workspace = true, optional = true }
+itertools = { workspace = true, optional = true }
 
 # `metrics` feature
 reth-metrics = { workspace = true, optional = true }
@@ -58,12 +55,11 @@ proptest.workspace = true
 proptest-arbitrary-interop.workspace = true
 
 [features]
-default = ["metrics", "async", "parallel"]
+default = ["metrics", "parallel"]
 metrics = ["reth-metrics", "dep:metrics", "reth-trie/metrics"]
-async = ["tokio/sync", "itertools"]
-parallel = ["rayon"]
+parallel = ["rayon", "itertools"]
 
 [[bench]]
 name = "root"
-required-features = ["async", "parallel"]
+required-features = ["parallel"]
 harness = false
