+++ crates/trie/parallel/benches/root.rs
@@ -13,7 +13,7 @@ use reth_trie::{
     TrieInput,
 };
 use reth_trie_db::{DatabaseHashedCursorFactory, DatabaseStateRoot};
-use reth_trie_parallel::{async_root::AsyncStateRoot, parallel_root::ParallelStateRoot};
+use reth_trie_parallel::parallel_root::ParallelStateRoot;
 use std::collections::HashMap;
 
 pub fn calculate_state_root(c: &mut Criterion) {
@@ -68,14 +68,6 @@ pub fn calculate_state_root(c: &mut Criterion) {
                     )
                 },
                 |calculator| async { calculator.incremental_root() },
-            );
-        });
-
-        // async root
-        group.bench_function(BenchmarkId::new("async root", size), |b| {
-            b.iter_with_setup(
-                || AsyncStateRoot::new(view.clone(), TrieInput::from_state(updated_state.clone())),
-                |calculator| calculator.incremental_root(),
             );
         });
     }
