+++ crates/trie/parallel/src/lib.rs
@@ -13,10 +13,6 @@ pub use storage_root_targets::StorageRootTargets;
 /// Parallel trie calculation stats.
 pub mod stats;
 
-/// Implementation of async state root computation.
-#[cfg(feature = "async")]
-pub mod async_root;
-
 /// Implementation of parallel state root computation.
 #[cfg(feature = "parallel")]
 pub mod parallel_root;
